+++ packages/discordx/examples/multiple-discord-instances/main.ts
@@ -28,7 +28,12 @@ export class Main {
     });
 
     await importx(dirname(import.meta.url) + "/commands/**/*.{js,ts}");
-    await this._client.login(process.env.BOT_TOKEN ?? "");
+
+    // let's start the bot
+    if (!process.env.BOT_TOKEN) {
+      throw Error("Could not find BOT_TOKEN in your environment");
+    }
+    await this._client.login(process.env.BOT_TOKEN);
   }
 }
