+++ packages/lava-player/examples/Main.ts
@@ -43,7 +43,12 @@ export class Main {
     });
 
     await importx(dirname(import.meta.url) + "/discords/**/*.{js,ts}");
-    await this._client.login(process.env.BOT_TOKEN ?? "");
+
+    // let's start the bot
+    if (!process.env.BOT_TOKEN) {
+      throw Error("Could not find BOT_TOKEN in your environment");
+    }
+    await this._client.login(process.env.BOT_TOKEN);
   }
 }
