+++ packages/utilities/examples/category/Main.ts
@@ -1,6 +1,6 @@
 import "reflect-metadata";
 
-import { importx } from "@discordx/importer";
+import { dirname, importx } from "@discordx/importer";
 import { Intents } from "discord.js";
 import { Client } from "discordx";
 
@@ -34,8 +34,13 @@ export class Main {
       this._client.executeInteraction(interaction);
     });
 
-    await importx(__dirname + "/discords/**/*.{js,ts}");
-    await this._client.login(process.env.BOT_TOKEN ?? "");
+    await importx(dirname(import.meta.url) + +"/discords/**/*.{js,ts}");
+
+    // let's start the bot
+    if (!process.env.BOT_TOKEN) {
+      throw Error("Could not find BOT_TOKEN in your environment");
+    }
+    await this._client.login(process.env.BOT_TOKEN);
   }
 }
