+++ packages/utilities/examples/category/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "compilerOptions": {
-    "module": "CommonJS",
+    "module": "ESNext",
     "target": "ESNext",
     "strict": true,
     "noImplicitAny": true,
