+++ packages/utilities/examples/guards/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "compilerOptions": {
-    "module": "CommonJS",
+    "module": "ESNext",
     "target": "ESNext",
     "strict": true,
     "noImplicitAny": true,
@@ -13,9 +13,5 @@
     "moduleResolution": "Node",
     "esModuleInterop": true
   },
-  "exclude": [
-    "node_modules",
-    "tests",
-    "examples"
-  ]
+  "exclude": ["node_modules", "tests", "examples"]
 }
