+++ packages/utilities/src/index.ts
@@ -1,5 +1,7 @@
 // category
 export * from "./category/index.js";
+
+// guard
 export * from "./guards/RateLimit.js";
-export * from "./utils/ITimedSet.js";
-export * from "./utils/TimedSet.js";
+export * from "./guards/utils/ITimedSet.js";
+export * from "./guards/utils/TimedSet.js";
